# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gerenciamento de Usuários e Restaurantes
  version: 1.0.0
  description: API para gerenciamento de usuários, restaurantes e itens do cardápio.
...
paths:
...
  /restaurantes:
    get:
      tags:
        - Restaurante
      summary: Buscar restaurantes por filtros
      description: Retorna uma lista de restaurantes de acordo com os filtros opcionais de nome, id do usuário e horário de funcionamento.
      operationId: buscarRestaurante
      parameters:
        - name: nome
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Nome'
        - name: usuarioId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UsuarioId'
        - name: horarioFuncionamento
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/HorarioFuncionamento'
      responses:
        '200':
          description: Lista de restaurantes encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurante'
        '204':
          description: Nenhum restaurante encontrado
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Restaurante
      summary: Criar novo restaurante
      description: Cria um novo restaurante com os dados fornecidos, incluindo o dono (usuário responsável).
      operationId: inserirRestaurante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestauranteRequest'
      responses:
        '201':
          description: Restaurante criado com sucesso
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Restaurante ou usuário dono já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /restaurantes/{id}:
    patch:
      tags:
        - Restaurante
      summary: Atualizar dados do restaurante
      description: Atualiza os dados de um restaurante existente identificado pelo nome do usuário do dono.
      operationId: atualizarRestaurante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestauranteRequest'
      responses:
        '200':
          description: Restaurante atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurante'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Restaurante não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Restaurante
      summary: Remover restaurante
      description: Remove um restaurante identificado pelo nome do usuário do dono.
      operationId: deletarRestaurante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '204':
          description: Restaurante removido com sucesso
        '404':
          description: Restaurante não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Id:
      type: integer
      format: int64
      example: 1
    UsuarioId:
      type: integer
      format: int64
      example: 1
...
    Descricao:
      type: string
      example: "Italiana"
    Nome:
      type: string
      example: "Sabor Caseiro"
    HorarioFuncionamento:
      type: string
      example: "10:00 - 22:00"
...
    Endereco:
      type: object
      properties:
        street:
          type: string
        number:
          type: string
        complement:
          type: string
        city:
          type: string
        state:
          type: string
      required:
        - street
        - number
        - city
        - state
...
    Restaurante:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        nome:
          $ref: '#/components/schemas/Nome'
        endereco:
          $ref: '#/components/schemas/Endereco'
        tipoCozinha:
          $ref: '#/components/schemas/Descricao'
        horarioFuncionamento:
          $ref: '#/components/schemas/HorarioFuncionamento'
        usuarioId:
          $ref: '#/components/schemas/UsuarioId'
      required:
        - nome
        - endereco
        - tipoCozinha
        - horarioFuncionamento
        - usuario
    RestauranteRequest:
      type: object
      properties:
        nome:
          $ref: '#/components/schemas/Nome'
        endereco:
          $ref: '#/components/schemas/Endereco'
        tipoCozinha:
          $ref: '#/components/schemas/Descricao'
        horarioFuncionamento:
          $ref: '#/components/schemas/HorarioFuncionamento'
      required:
        - nome
        - endereco
        - tipoCozinha
        - horarioFuncionamento
        - usuario
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-05-14T15:42:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Recurso não encontrado"
        path:
          type: string
          example: "/restaurantes/restaurante123/item-cardapio"